# contracts/openapi/v1.yaml
openapi: 3.1.0
info:
  title: Policy Orchestrator & Algorithm Execution API
  version: 1.0.0
  description: >-
    Enterprise-grade, vendor-agnostic REST API for repository invariant auditing,
    grounded hybrid GraphRAG, autonomous AI Agent orchestration, and direct execution
    of all 152 Layer 1 algorithms across Search, Observability, Update, Vector, Graph, Filter, and Transform categories.
    Strictly conforms to the RFC API Request & Response Structure Specification (v5.0).
servers:
  - url: http://localhost:8000
    description: Local development orchestrator
paths:
...
  /api/v1/algos/observability/outline:
    post:
      summary: Hierarchical Symbol Outline Generation (Observability Algorithm Engine)
      operationId: generateFileOutline
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [file_path]
              properties:
                file_path:
                  type: string
      responses:
        '200':
          description: Code symbols hierarchy and markdown outline
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_AlgoResponse'
  /api/v1/algos/observability/dependencies:
    post:
      summary: Module Import Graph & Cycle Detection (Observability Algorithm Engine)
      operationId: analyzeDependencies
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                directory:
                  type: string
                  default: "."
      responses:
        '200':
          description: Import DAG report and cycle detection findings
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_AlgoResponse'
  /api/v1/algos/observability/lint-comments:
    post:
      summary: Zero-Inline-Comment Doctrine Linter (Observability Algorithm Engine)
      operationId: lintComments
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [file_path]
              properties:
                file_path:
                  type: string
      responses:
        '200':
          description: Zero-inline comment doctrine compliance report
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_AlgoResponse'
...
  /api/v1/algos/graph/bfs:
    post:
      summary: Breadth-First Search Traversal (ALGO-GRAPH-01)
      operationId: graphBfs
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [adjacency_list, start_node]
              properties:
                adjacency_list:
                  type: object
                  additionalProperties:
                    type: array
                    items: { type: string }
                start_node: { type: string }
                max_depth: { type: integer, default: -1 }
      responses:
        '200':
          description: Visited order and distances
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_AlgoResponse'
  /api/v1/algos/graph/dfs:
    post:
      summary: Depth-First Search Traversal (ALGO-GRAPH-02)
      operationId: graphDfs
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [adjacency_list, start_node]
              properties:
                adjacency_list:
                  type: object
                  additionalProperties:
                    type: array
                    items: { type: string }
                start_node: { type: string }
                max_depth: { type: integer, default: -1 }
      responses:
        '200':
          description: Visited order and cycle detection
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_AlgoResponse'
  /api/v1/algos/graph/dijkstra:
    post:
      summary: Dijkstra Shortest Path (ALGO-GRAPH-03)
      operationId: graphDijkstra
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [weighted_edges, start_node]
              properties:
                weighted_edges:
                  type: array
                  items:
                    type: object
                    required: [source, target, weight]
                    properties:
                      source: { type: string }
                      target: { type: string }
                      weight: { type: number }
                start_node: { type: string }
                target_node: { type: string }
      responses:
        '200':
          description: Shortest path distances and reconstructed paths
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_AlgoResponse'
  /api/v1/algos/graph/astar:
    post:
      summary: A* Heuristic Search (ALGO-GRAPH-04)
      operationId: graphAstar
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [weighted_edges, start_node, target_node]
              properties:
                weighted_edges:
                  type: array
                  items:
                    type: object
                    required: [source, target, weight]
                    properties:
                      source: { type: string }
                      target: { type: string }
                      weight: { type: number }
                start_node: { type: string }
                target_node: { type: string }
                heuristics:
                  type: object
                  additionalProperties: { type: number }
      responses:
        '200':
          description: Optimal path and search cost
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_AlgoResponse'
  /api/v1/algos/graph/pagerank:
    post:
      summary: PageRank Centrality (ALGO-GRAPH-05)
      operationId: graphPageRank
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [adjacency_list]
              properties:
                adjacency_list:
                  type: object
                  additionalProperties:
                    type: array
                    items: { type: string }
                damping_factor: { type: number, default: 0.85 }
                max_iterations: { type: integer, default: 100 }
                tolerance: { type: number, default: 1e-6 }
      responses:
        '200':
          description: PageRank centrality scores
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_AlgoResponse'
  /api/v1/algos/graph/degree-centrality:
    post:
      summary: Degree Centrality (ALGO-GRAPH-06)
      operationId: graphDegreeCentrality
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [adjacency_list]
              properties:
                adjacency_list:
                  type: object
                  additionalProperties:
                    type: array
                    items: { type: string }
                normalized: { type: boolean, default: true }
      responses:
        '200':
          description: In-degree, out-degree and total degree centrality
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_AlgoResponse'
  /api/v1/algos/graph/connected-components:
    post:
      summary: Connected Components Union-Find (ALGO-GRAPH-07)
      operationId: graphConnectedComponents
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [edges]
              properties:
                edges:
                  type: array
                  items:
                    type: array
                    items: { type: string }
                nodes:
                  type: array
                  items: { type: string }
      responses:
        '200':
          description: Partitioned connected components
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_AlgoResponse'
  /api/v1/algos/graph/tarjan-scc:
    post:
      summary: Tarjan Strongly Connected Components (ALGO-GRAPH-08)
      operationId: graphTarjanScc
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [adjacency_list]
              properties:
                adjacency_list:
                  type: object
                  additionalProperties:
                    type: array
                    items: { type: string }
      responses:
        '200':
          description: List of strongly connected components
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_AlgoResponse'
  /api/v1/algos/graph/subgraph-match:
    post:
      summary: Subgraph Pattern Matching VF2 (ALGO-GRAPH-09)
      operationId: graphSubgraphMatch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [target_graph, pattern_graph]
              properties:
                target_graph:
                  type: object
                  additionalProperties:
                    type: array
                    items: { type: string }
                pattern_graph:
                  type: object
                  additionalProperties:
                    type: array
                    items: { type: string }
                max_matches: { type: integer, default: 100 }
      responses:
        '200':
          description: Subgraph pattern matching solutions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_AlgoResponse'
...
components:
  schemas:
    ResponseMeta:
      type: object
      required:
        - requestId
        - correlationId
        - timestamp
        - apiVersion
        - status
        - trace_id
      properties:
        requestId:
          type: string
          example: "req-1700000000000-a1b2c3"
        correlationId:
          type: string
          example: "corr-1700000000000-x9y8z7"
        causationId:
          type: string
          nullable: true
        timestamp:
          type: string
          example: "2026-10-05T14:00:00.000Z"
        executionTimeMs:
          type: integer
          example: 12
        apiVersion:
          type: string
          example: "v1"
        status:
          type: string
          enum: [success, error]
        trace_id:
          type: string
          example: "0af7651916cd43dd8448eb211c80319c"
        auditLogId:
          type: string
          nullable: true
        operationId:
          type: string
          nullable: true
        deprecatedFields:
          type: array
          items: { type: string }
          nullable: true
        supportedVersions:
          type: array
          items: { type: string }
        staleness:
          type: object
          nullable: true
        pagination:
          type: object
          nullable: true
        links:
          type: object
          nullable: true
    ErrorContainer:
      type: object
      required:
        - code
        - message
        - retryable
      properties:
        code:
          type: string
          example: "VALIDATION_FAILED"
        message:
          type: string
          example: "One or more payload validation checks failed."
        retryable:
          type: boolean
          default: false
        details:
          type: array
          items:
            type: object
    ErrorItem:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: "VALIDATION_FAILED"
        message:
          type: string
          example: "One or more payload validation checks failed."
        details:
          type: object
...
    Envelope_AlgoResponse:
      type: object
      required: [success, statusCode, meta, data]
      properties:
        success: { type: boolean, example: true }
        statusCode: { type: integer, example: 200 }
        data: { type: object }
        meta: { $ref: '#/components/schemas/ResponseMeta' }
        error: { $ref: '#/components/schemas/ErrorContainer' }
        errors: { type: array, items: { $ref: '#/components/schemas/ErrorItem' }, nullable: true }
...